+++ pycomex/cli.py
@@ -8,7 +8,7 @@ from pycomex.util import get_version
 @click.group(invoke_without_command=True)
 @click.option("-v", "--version", is_flag=True)
 @click.pass_context
-def main(ctx: click.Context, version: bool):
+def cli(ctx: click.Context, version: bool):
     """Console script for pycomex."""
     if version:
         version = get_version()
@@ -17,4 +17,4 @@ def main(ctx: click.Context, version: bool):
 
 
 if __name__ == "__main__":
-    main()  # pragma: no cover
+    cli()  # pragma: no cover